fix: add tests for create-sessions workflow [#5]
Here we add the test suite for creating a session via workflow and then
fix up the small issues we discovered during the testing. Like the fact
that Mojolicious happily turns our array ref into a scalar if there's
only one element in it.

We also went through and move everything over from feature 'class' to
Object::Pad for now.

And I silenced a warning that cropped up with builtin::export_lexically