feat: continuations [#5]
We need to be able to pause a workflow (session_creation) and start a new
workflow (event_creation) and have the results of the new workflow feed
back into the original workflow.

This is kinda the definition of a [continuation][1].

You can add a continuation_id to the workflow run and when the workflow
ends it will look and if there's a continuation_id it'll push it's data
onto to that run as well as writing it out.

[1]: https://en.wikipedia.org/wiki/Continuation